+++ src/Services/CRM/Deal/Service/Deal.php
@@ -236,7 +236,7 @@ class Deal extends AbstractService
      * @throws BaseException
      * @throws TransportException
      */
-    public function update(int $id, array $fields, array $params): UpdatedItemResult
+    public function update(int $id, array $fields, array $params = []): UpdatedItemResult
     {
         return new UpdatedItemResult(
             $this->core->call(