+++ src/Services/CRM/Userfield/Result/UserfieldTypeItemResult.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\CRM\Userfield\Result;
+
+use Bitrix24\SDK\Core\Result\AbstractItem;
+
+/**
+ * @property-read string $ID
+ * @property-read string $title
+ */
+class UserfieldTypeItemResult extends AbstractItem
+{
+}